feat: update the useFetchUniList hook to use separate constants and types file so I can export them to the formatCSV util
