+++ src/apis/constants.js
@@ -1,0 +1,6 @@
+export const FETCH_UNI_ACTION_TYPES = {
+  ERROR: 'error',
+  LOADING: 'loading',
+  RESET: 'reset',
+  SET: 'set',
+};
